pyproject.toml
[tool.poetry]
name = "py-cron"
version = "0.1.0"
description = ""
authors = ["eziox <ezioxxk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
anthropic = "^0.32.0"
sqlalchemy = "^2.0.31"
apscheduler = "^3.10.4"
prometheus-client = "^0.20.0"
instructor = "^1.3.7"
toolz = "^0.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
